parannuksia & ts build
