+++ package.json
@@ -5,8 +5,10 @@
     "main": "index.js",
     "scripts": {
         "test": "jest",
-        "start": "node ./src/index.js",
-        "dev": "nodemon ./src/index.ts"
+        "start-ts": "ts-node ./src/index.ts",
+        "start": "node ./dist/index.js",
+        "dev": "nodemon ./src/index.ts",
+        "build": "tsc -p ."
     },
     "keywords": [],
     "author": "",
